Add basic navigation between info and waiting screen
